requirements.txt
twikit==2.3.1
google-generativeai>=0.8.0
openpyxl>=3.1.0
